[add]ファイル分けました、注意画面
